chore: tidy up folder structure
