+++ src/components/widgets/plotly.tsx
